Split native runtime bridge helpers
